requirements.txt
"""
Archivo de requisitos para el Sistema de Documentación Automatizada

Este archivo lista todas las dependencias necesarias para ejecutar el sistema.
"""

# Dependencias básicas
flask==2.3.3
pillow==10.0.0
pyautogui==0.9.54
pynput==1.7.6
opencv-python==4.8.0.76
pytesseract==0.3.10
jinja2==3.1.2
markdown==3.4.4
weasyprint==59.0

# Dependencias para modelos multimodales
torch==2.0.1
transformers==4.33.2
accelerate==0.23.0
bitsandbytes==0.41.1

# Dependencias para desarrollo y pruebas
pytest==7.4.0
pytest-flask==1.2.0
